==> notebooks/1-DATA_EDA_AVT.ipynb <==
Datos extraídos mediante webscraping de **[FlowAgility](https://flowagility.com)**

👉 *[FlowAgility-webscraper](https://github.com/sigmasoldi3r/flowagility-webscraper.git)*

''' 

En este notebook hago la primera visualización del JSON que el programa anterior me ha realizado, 
estudio como extraer la información que necesito de las columnas "Runs" e "Info".

--> Consigo extraer la información del Suelo y Techo de cada pista :)

'''

import pandas as pd
import numpy as np
from bs4 import BeautifulSoup

df_agility = pd.read_json("./data/agility-data.json")
df_agility.dropna(axis=0, inplace=True)
df_agility.info()

df_agility.head(20)

df_agility["indice"] = lista
df_agility = df_agility.set_index("indice")

lista = list(range(682))
lista_len = []
for cantidad in lista:
    lista_len = np.append(lista_len, len(df_agility["runs"][cantidad]))

df_agility["len_runs"] = lista_len

df_agility["len_runs"]

df_agility["runs"][13][1]

# Columnas del new dataset para investigar mejor:

- Día (runs --> "type")
- Title ✔️
- Techo (info --> "event"-"conditions"-"Roof") ✔️
- Suelo (info --> "event"-"conditions"-"Floor") ✔️
- Manga (runs --> "results"[0]= Agility, "results[1]= Jumping ?)
- (runs --> "results[2]= Binominal Info: )
    - License
    - Breed
    - Age
    - Gender
    - Hight_cm
    - Cualification
    - Speed
    - Time
    - Faults
    - Refusals


prueva = np.array(df_agility["runs"][0])
dic_prueva = {}
for titulo, texto in prueva[0].items():
    print(titulo)
    if titulo == "type":
        #print(texto)
        dic_prueva[titulo] = texto

df_prueva = pd.DataFrame([dic_prueva])
df_prueva


len(prueva)

df_agility["info"][0]

# INFO --> Techo y Suelo
lista = list(range(682))
lista_roof = []
lista_floor = []
for cantidad in lista:
    for clave, valor in df_agility["info"][cantidad].items():
        if clave == "dataTable":
            lista_floor.append(valor['Conditions']["Floor"])
            lista_roof.append(valor['Conditions']["Roof"])

df_agility["Roof"] = lista_roof
df_agility["Floor"] = lista_floor

df_agility_2 = df_agility.copy()

df_agility.drop(["issuer", "flag","participantsUrl","runsUrl","infoUrl","isCancelled","unknownUrls"], axis=1, inplace=True)

df_agility.to_json("df_agility.json", orient="records")
