==> notebooks/3-DEFDATA_EDA_AVT.ipynb <==
''' 

En este notebook Junto la información del dataset realizado en el notebook número 1 y 2.

También hay alguna tarea de limpieza de datos que no nos interesan para el estudio. 

'''

import pandas as pd
import numpy as np

df_agility = pd.read_json("./data/df_agility.json")
data_runs = pd.read_json("./data/data_runs.json")

df_agility.head(1)

data_runs.head(1)

dic_join = {
    "Competition": [],
    "Location": [],
    "Roof": [],
    "Floor": []
}


for i in data_runs["Index_data"]:
    dic_join["Competition"].append(df_agility["title"][i])
    dic_join["Location"].append(df_agility["location"][i])
    dic_join["Roof"].append(df_agility["Roof"][i])
    dic_join["Floor"].append(df_agility["Floor"][i])

data_runs["Competition"] = dic_join["Competition"]
data_runs["Location"] = dic_join["Location"]
data_runs["Roof"] = dic_join["Roof"]
data_runs["Floor"] = dic_join["Floor"]

data_runs.info()

data_runs["License"].value_counts()

# Eliminamos los pre-agility y los perros que no tienen Licencia.
for licencia in data_runs["License"]:
    if licencia == "PRE":
        indices_a_eliminar_1 = (data_runs["License"].index[data_runs["License"] == licencia].to_list())
    if licencia == "-":
        indices_a_eliminar_2 = (data_runs["License"].index[data_runs["License"] == licencia].to_list())

len(indices_a_eliminar_1)

len(indices_a_eliminar_2)

#ELIMINACIÓN FINAL:
data_runs_2 = data_runs.copy()

data_runs = data_runs.drop(indices_a_eliminar_1)
data_runs = data_runs.drop(indices_a_eliminar_2)

data_runs.info()

eliminadas = 35572 - 1282 - 1496
eliminadas

data_runs.info()

data_runs.to_json("agility_runs_df.json", orient="records")


data_runs.to_csv("./data/agility_runs_df.csv")
