==> notebooks/4-LIMPIEZA_EDA_AVT.ipynb <==
'''
En este notebook limpio y modifico algunos datos como las columnas "Speed" y "Time" que son object

--> Columnas puestas como variantes numéricas (float64)

'''

import numpy as np
import pandas as pd

df_agility = pd.read_csv("./data/agility_runs_df.csv")

# Descartamos competiciones en Mexico que no tienen nº Licencia:
df_agility.loc[df_agility["License"].isna()]

df_agility.dropna(axis="index", inplace=True)
df_agility.info()

# 2753 participantes 
df_agility["License"].describe()

# 620 competiciones
# 32771 pistas
df_agility["Competition"].describe()

df_agility

df_agility.drop(columns = ["Unnamed: 0", "Indice"], inplace = True)

df_agility["Speed"]

def corrige_speed(velocidad):
    nueva_velocidad = float(velocidad.replace(" m/s",""))
    return float(nueva_velocidad)

def speed_0(velocidad):
    eliminados = float(velocidad.replace("-", "0"))
    return float(eliminados)

def corrige_tiempo(tiempo):
    nuevo_tiempo = float(tiempo.replace(" s", ""))
    return float(nuevo_tiempo)

v_elim = df_agility["Speed"].isin(["-"])
v_elim
    

s_elim = df_agility["Time (s)"].isin(["-"])
s_elim

df_agility["Speed"].loc[v_elim] = df_agility["Speed"].loc[v_elim].apply(speed_0)


df_agility["Time (s)"].loc[s_elim] = df_agility["Time (s)"].loc[s_elim].apply(speed_0)

df_agility["Speed"].loc[df_agility["Speed"] != 0] = df_agility["Speed"].loc[df_agility["Speed"] != 0].apply(corrige_speed)

df_agility["Time (s)"].loc[df_agility["Time (s)"] != 0] = df_agility["Time (s)"].loc[df_agility["Time (s)"] != 0].apply(corrige_tiempo)

df_agility = df_agility.rename(columns ={"Speed":"Speed m/s", "Time":"Time (s)"})
df_agility

df_agility["Speed m/s"] = df_agility["Speed m/s"].astype(float)
df_agility["Time (s)"] = df_agility["Time (s)"].astype(float)

df_agility.info()

df_agility.to_csv("5-DATA_AGILITY.csv")
