==> notebooks/5-ORDEN_EDA_AVT.ipynb <==
'''
En este notebook transformo la columna "Qualification" en la columna "Eliminados". Elimino los No Presentados.
Rehago tambien la columna "Manga", ahora "Pista".

--> Columnas "Faults" y "Refusals" puestas como variantes numéricas (int32)

'''

import numpy as np
import pandas as pd

df_agility = pd.read_csv("./data/5-DATA_AGILITY.csv")
primer_df = pd.read_json("./data/df_agility.json")

df_agility.head(5)

df_agility.drop(columns = ["Unnamed: 0"], inplace = True)

df_agility["Qualification"].value_counts()

eliminados = []
for elim in df_agility["Qualification"]:
    if elim == "DISQ" or elim == "DISQ (R)" or elim == "DISQ (T)":
        eliminados.append("ELIM")
    else:
        eliminados.append(elim)

df_agility["Eliminados"] = eliminados

df_agility["Eliminados"].value_counts()

# Eliminamos los participantes NP (No Presentados) y NQ (No tienen Qualificación ni resultado de Classificación)
index_NP = df_agility.loc[df_agility["Eliminados"] == "NQ"].index.to_list()

df_agility = df_agility.drop(index_NP)
df_agility.info()

indice = list(range(32611))
df_agility["indice"] = indice
df_agility = df_agility.set_index("indice")

df_agility.info()

| Columna | Descripción| Comentario|
|---|---|---|
|Index_data| Número relacionado con el evento de la competición| generado en el primero dataset|
|Fecha| Día de la competición | + Si no tiene año es del 2024 + modificar formato|
|Manga| Tipo de pista| Jumping o Agility|
|License| identificador individual del Binomio| Groupby con Breed, Gender, Hight_cm|
|Breed| Razas de los perros|-|
|Gender| Hembra o Macho|-|
|Hight_cm| Altura de la cruz|-|
|Classification| 1r, 2o, 3o| Ganadores de la competición|
|Qualification| EXC_0,DISQ,EXC,V_GOOD,GOOD,DISQ (R),NP,NQ,DISQ (T)| EXC_0 = Pista a cero, DISQ (R y T?)= Eliminado, NP= No presentado|
|Speed| Velocidad del perro|-|
|Time| Tiempo en pista|-|
|Faults| nº de Faltas realizadas en pista|-|
|Refusals| nº de Faltas leves realizadas en pista|-|
|Competition| Nombre de la competición| relacionado con Index_data|
|Location| Lugar de la competicion| relacionado con Competición|
|Roof| Tipo de techo de la competición| Outdoor o indoor|
|Floor| Tipo de suelo| ...|
|Eliminados| Misma columna que Qualification pero limpia|-|



# Borro la Columna Qualification porque ya tenemos la de Eliminados que nos da la misma información limpia
del df_agility["Qualification"]

# Rehago la fecha:
list_date = []
for indice in df_agility["Index_data"]:
    list_date.append(primer_df["shortDate"][indice])

df_agility["Fecha_Competición"] = list_date

# Decido no canviar a formato datetime. Demasiados patrones distintos y no es tan relevante para el estudio.
df_agility["Fecha_Competición"].unique()

df_agility["Manga"].value_counts()

lista_pista = []
for manga in df_agility["Manga"]:
    if manga.startswith("Jumping"):
        lista_pista.append("Jumping")
    if manga.startswith("Agility"):
        lista_pista.append("Agility")

df_agility["Pista"] = lista_pista
df_agility["Pista"].value_counts()

# Elimino la columna Manga porque ya se ha creado Pista que está limpia
del df_agility["Manga"]

df_agility.info()

def corrige_a_cero(resultado):
    a_cero = float(resultado.replace("-", "0"))
    return float(a_cero)

def corrige_eliminado(resultado):
    eliminado = float(resultado.replace("-", "100"))
    return float(eliminado)

# Marcamos los 0 reales y los eliminados en Faltas y Rehuses
es_eliminado = df_agility["Eliminados"] == "ELIM"
faltas_elim = df_agility["Faults"].isin(["-"])
filtro = es_eliminado & faltas_elim

es_eliminado2 = df_agility["Eliminados"] == "ELIM"
faltas_elim2 = df_agility["Refusals"].isin(["-"])
mascara = es_eliminado2 & faltas_elim2

df_agility["Faults"].loc[filtro] = df_agility["Faults"].loc[filtro].apply(corrige_eliminado)

df_agility["Refusals"].loc[mascara] = df_agility["Refusals"].loc[mascara].apply(corrige_eliminado)

ceros_reales = df_agility["Faults"].isin(["-"])
df_agility["Faults"].loc[ceros_reales] = df_agility["Faults"].loc[ceros_reales].apply(corrige_a_cero)

ceros_reales2 = df_agility["Refusals"].isin(["-"])
df_agility["Refusals"].loc[ceros_reales2] = df_agility["Refusals"].loc[ceros_reales2].apply(corrige_a_cero)

df_agility["Faults"] = df_agility["Faults"].astype(int)
df_agility["Refusals"] =  df_agility["Refusals"].astype(int)

df_agility.info()

df_agility.to_csv("6_DATA_AGILITY.csv")
