==> notebooks/6_CATEGORIAS_EDA_AVT.ipynb <==
'''
En este notebook se crea la columna "Categoria"

'''

import numpy as np
import pandas as pd

Dividimos el estudio en las distintas categorias según normativa RSCE: https://www.rsce.es/images/rsce/Agility/Reglamentos/REGLAMENTO_AGILITY_ENERO_2024_VIGOR.pdf

Artículo 15.- Se prevén cuatro (4) categorías en competiciones y pruebas de Agility nacionales que organice
o autorice organizar la RSCE y cinco (5) alturas de salto:

-  La categoría pequeña (SMALL), para perros por debajo de 35 cm de altura a la cruz, subdividida en
dos subcategorías, cada una con su altura de salto:
    - La subcategoría XS para perros por debajo de 28 cm (27,99 cm) de altura a la cruz, con una altura
de salto de 15 a 20 cm.
    - La subcategoría S para perros de 28 hasta 35 cm (34,99 cm) de altura a la cruz, con una altura de
salto de 25 a 30 cm.
- La categoría mediana (MEDIUM), para perros de 35 hasta 43 cm (42,99 cm) de altura a la cruz, con
una altura de salto de 35 a 40 cm.
- La categoría intermedia (INTERMEDIATE), para perros de 43 hasta 48 cm (47,99 cm) de altura a la
cruz, con una altura de salto de 45 a 50 cm.
- La categoría estándar (LARGE), para perros de 48 cm, o más, de altura a la cruz, con una altura de
salto de 55 a 60 cm.

df_agility = pd.read_csv("./data/6_DATA_AGILITY.csv")

df_agility.info()

categoria = []
for medida in df_agility["Hight_cm"]:
    if medida < 28:
        categoria.append("XS")
    if medida >= 28 and medida < 35:
        categoria.append("S")
    if medida >= 35 and medida < 43:
        categoria.append("M")
    if medida >= 43 and medida < 48:
        categoria.append("I")
    if medida >= 48:
        categoria.append("L")


df_agility["Categoria"] = categoria

df_agility["Categoria"].value_counts()

df_agility.to_csv("6_DATA_AGILITY_CATEGORIAS.csv")
